+++ codex-rs/app-server/tests/suite/v2/mcp_resource.rs
@@ -216,6 +216,7 @@ async fn mcp_resource_read_returns_error_for_unknown_thread() -> Result<()> {
             },
             capabilities: None,
         },
+        identity_key: None,
         channel_capacity: in_process::DEFAULT_IN_PROCESS_CHANNEL_CAPACITY,
     })
     .await?;
