+++ codex-rs/thread-store/src/remote/list_threads.rs
@@ -272,10 +272,13 @@ mod tests {
                 page_size: 2,
                 cursor: Some("cursor-1".to_string()),
                 sort_key: ThreadSortKey::UpdatedAt,
+                sort_direction: crate::SortDirection::Desc,
                 allowed_sources: vec![SessionSource::Cli],
                 model_providers: Some(vec!["openai".to_string()]),
+                cwd_filters: Some(vec![PathBuf::from("/workspace")]),
                 archived: true,
                 search_term: Some("needle".to_string()),
+                use_state_db_only: true,
             })
             .await
             .expect("list threads");
